refactor: isolate session revoke action
