+++ packages/ui/src/actions/revokeSession.d.ts
@@ -1,0 +1,5 @@
+export declare function revoke(id: string): Promise<{
+    success: boolean;
+    error?: string;
+}>;
+//# sourceMappingURL=revokeSession.d.ts.map
