+++ packages/ui/src/components/account/index.ts
@@ -3,8 +3,5 @@ export { default as ProfilePage, metadata as profileMetadata } from "./Profile";
 export { default as OrdersPage, metadata as ordersMetadata } from "./Orders";
 export { default as MfaSetup } from "./MfaSetup";
 export { default as MfaChallenge } from "./MfaChallenge";
-export {
-  default as SessionsPage,
-  metadata as sessionsMetadata,
-  revoke,
-} from "./Sessions";
+export { default as SessionsPage, metadata as sessionsMetadata } from "./Sessions";
+export { revoke } from "../../actions/revokeSession";
